requirements.txt
Flask==1.1.2
firebase_admin==4.5.1
Pillow==8.1.0
